add PartialDate; update and refactor archaeology dtos
