+++ src/fragmentarium/application/FragmentService.ts
@@ -25,7 +25,7 @@ import { ManuscriptAttestation } from 'corpus/domain/manuscriptAttestation'
 import { FragmentQuery } from 'query/FragmentQuery'
 import { MesopotamianDate } from 'chronology/domain/Date'
 import { FragmentAfoRegisterQueryResult, QueryResult } from 'query/QueryResult'
-import { ArchaeologyDto } from 'fragmentarium/domain/archaeology'
+import { ArchaeologyDto } from 'fragmentarium/domain/archaeologyDtos'
 
 export const onError = (error) => {
   if (error.message === '403 Forbidden') {
