+++ src/fragmentarium/domain/archaeology.ts
@@ -1,10 +1,7 @@
 import Reference from 'bibliography/domain/Reference'
-import MuseumNumber, { museumNumberToString } from './MuseumNumber'
 import { Provenances } from 'corpus/domain/provenance'
 import _ from 'lodash'
 import { immerable } from 'immer'
-import { ReferenceDto } from 'bibliography/domain/referenceDto'
-import createReference from 'bibliography/application/createReference'
 
 export const excavationSites = {
   ..._.omit(Provenances, 'Standard Text'),
@@ -29,27 +26,39 @@ export interface ExcavationPlan {
   readonly svg: string
   readonly references: readonly Reference[]
 }
-export type CommentedDateRange = {
-  start?: number
-  end?: number
-  notes?: string
-}
-export type CommentedDateRangeDto = {
-  start?: number
-  end?: number
-  notes?: string
-}
+export class PartialDate {
+  readonly year: number
+  readonly month?: number | null
+  readonly day?: number | null
 
-function makeDate(date?: number) {
-  return date || date === 0 ? `${Math.abs(date)}${date < 0 ? ' BCE' : ''}` : ''
+  constructor(
+    year: number,
+    month: number | null = null,
+    day: number | null = null
+  ) {
+    this.year = year
+    this.month = month
+    this.day = day
+  }
+
+  toString(): string {
+    return this.year >= 0
+      ? _.reject([this.year, this.month, this.day], _.isNil).join('/')
+      : `${Math.abs(this.year)} BCE`
+  }
 }
-function pad(s?: string, left = ' ', right = ' '): string {
+export type CommentedDateRange = {
+  start: PartialDate
+  end?: PartialDate | null
+  notes?: string | null
+}
+function pad(s?: string | number | null, left = ' ', right = ' '): string {
   return s ? `${left}${s}${right}` : ''
 }
-function padLeft(s?: string, left = ' '): string {
+function padLeft(s?: string | number | null, left = ' '): string {
   return pad(s, left, '')
 }
-function padRight(s: string, right = ' '): string {
+function padRight(s: string | number | null, right = ' '): string {
   return pad(s, '', right)
 }
 
@@ -72,8 +81,8 @@ export class Findspot {
   ) {}
 
   private dateString(): string {
-    const start = makeDate(this.dateRange?.start)
-    const end = makeDate(this.dateRange?.end)
+    const start = this.dateRange?.start.toString()
+    const end = this.dateRange?.end?.toString()
     const notes = padLeft(this.dateRange?.notes, ', ')
 
     return end ? ` (${start} - ${end}${notes})` : start ? ` (${start})` : ''
@@ -102,104 +111,4 @@ export interface Archaeology {
   readonly isRegularExcavation?: boolean
   readonly findspotId?: number | null
   readonly findspot?: Findspot | null
-}
-
-interface PlanDto {
-  svg: string
-  references: readonly ReferenceDto[]
-}
-
-export type FindspotDto = Pick<
-  Findspot,
-  | 'area'
-  | 'building'
-  | 'buildingType'
-  | 'levelLayerPhase'
-  | 'room'
-  | 'context'
-  | 'primaryContext'
-  | 'notes'
-> & {
-  _id: number
-  site: SiteKey
-  dateRange: CommentedDateRangeDto | null
-  plans: readonly PlanDto[]
-}
-
-export type ArchaeologyDto = Omit<Archaeology, 'site' | 'findspot'> & {
-  site?: SiteKey
-  findspot?: FindspotDto | null
-}
-
-export function fromPlanDto(dto: PlanDto): ExcavationPlan {
-  return {
-    svg: dto.svg,
-    references: dto.references.map(createReference),
-  }
-}
-export function toPlanDto(plan: ExcavationPlan): PlanDto {
-  return {
-    svg: plan.svg,
-    references: plan.references.map((reference) => ({
-      ..._.pick(reference, 'id', 'type', 'pages', 'notes', 'linesCited'),
-      document: reference.document.toCslData(),
-    })),
-  }
-}
-
-export function fromFindspotDto(dto: FindspotDto): Findspot {
-  return new Findspot(
-    dto._id,
-    excavationSites[dto.site || ''],
-    dto.area,
-    dto.building,
-    dto.buildingType,
-    dto.levelLayerPhase,
-    dto.dateRange,
-    dto.plans.map(fromPlanDto),
-    dto.room,
-    dto.context,
-    dto.primaryContext,
-    dto.notes
-  )
-}
-
-export function toFindspotDto(findspot: Findspot): FindspotDto {
-  return {
-    _id: findspot.id,
-    area: findspot.area,
-    building: findspot.building,
-    buildingType: findspot.buildingType,
-    levelLayerPhase: findspot.levelLayerPhase,
-    room: findspot.room,
-    context: findspot.context,
-    primaryContext: findspot.primaryContext,
-    notes: findspot.notes,
-    site: findspot.site.name as SiteKey,
-    dateRange: findspot.dateRange,
-    plans: findspot.plans.map(toPlanDto),
-  }
-}
-
-export function createArchaeology(
-  dto: Omit<ArchaeologyDto, 'excavationNumber'> & {
-    excavationNumber?: MuseumNumber
-  }
-): Archaeology {
-  return {
-    ...dto,
-    excavationNumber: dto.excavationNumber
-      ? museumNumberToString(dto.excavationNumber)
-      : undefined,
-    site: excavationSites[dto.site || ''],
-    findspot: dto.findspot ? fromFindspotDto(dto.findspot) : null,
-  }
-}
-
-export function toArchaeologyDto(archaeology: Archaeology): ArchaeologyDto {
-  return {
-    ...archaeology,
-    site: (archaeology.site?.name || '') as SiteKey,
-    findspot: archaeology.findspot ? toFindspotDto(archaeology.findspot) : null,
-  }
 }
