+++ src/fragmentarium/ui/fragment/ArchaeologyEditor.tsx
@@ -4,12 +4,14 @@ import { Form, Col, Button } from 'react-bootstrap'
 import Select, { ValueType } from 'react-select'
 import {
   Archaeology,
-  ArchaeologyDto,
   Findspot,
   SiteKey,
   excavationSites,
-  toFindspotDto,
 } from 'fragmentarium/domain/archaeology'
+import {
+  ArchaeologyDto,
+  toFindspotDto,
+} from 'fragmentarium/domain/archaeologyDtos'
 import { Fragment } from 'fragmentarium/domain/fragment'
 import withData from 'http/withData'
 import { FindspotService } from 'fragmentarium/application/FindspotService'
